+++ Locket/domain/src/main/java/com/ssafy/locket/model/payment_history/PaymentDailyHistoryItem.kt
@@ -2,11 +2,12 @@ package com.ssafy.locket.model.payment_history
 
 import com.ssafy.locket.model.base.BaseModel
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 
 @Parcelize
 data class PaymentDailyHistoryItem(
     val cardName: String,
     val category: String,
     val storeName: String,
-    val totalAmount: Int
+    val totalAmount: BigDecimal
 ): BaseModel
